+++ lib/src/video_player/video_player.dart
@@ -531,6 +531,14 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
   /// If [moment] is outside of the video's full range it will be automatically
   /// and silently clamped.
   Future<void> seekTo(Duration? position) async {
+    bool isPlaying = value.isPlaying;
+    final int positionInMs = value.position.inMilliseconds;
+    final int durationInMs = value.duration?.inMilliseconds ?? 0;
+
+    if (positionInMs >= durationInMs && position?.inMilliseconds == 0) {
+      isPlaying = true;
+    }
+
     if (_isDisposed) {
       return;
     }
@@ -545,6 +553,11 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
     }
     await _videoPlayerPlatform.seekTo(_textureId, positionToSeek);
     _updatePosition(position);
+    if (isPlaying) {
+      play();
+    } else {
+      pause();
+    }
   }
 
   /// Sets the audio volume of [this].
@@ -560,8 +573,14 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
   ///
   /// [speed] indicates a value between 0.0 and 2.0 on a linear scale.
   Future<void> setSpeed(double speed) async {
-    value = value.copyWith(speed: speed);
-    await _applySpeed();
+    final double previousSpeed = value.speed;
+    try {
+      value = value.copyWith(speed: speed);
+      await _applySpeed();
+    } catch (exception) {
+      value = value.copyWith(speed: previousSpeed);
+      rethrow;
+    }
   }
 
   /// Sets the video track parameters of [this]